Update asteroid generation, add background
Asteroid generation has been improved to take averages of larger blocks,
and only place up to maximum properly. It also uses better simplex noise
settings to result in a more "playable" map.

A new background using a noise texture and a shader to create a
starfield has been added.

UI updated to allow for regeneration of the map, and saving of the seed.